operator/requirements.txt
aiohttp==3.7.3
aiojobs==0.3.0
async-timeout==3.0.1
attrs==20.3.0
cachetools==4.1.1
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
google-auth==1.23.0
idna==2.10
iso8601==0.1.13
jsonschema==3.2.0
kopf==0.28.1
kubernetes==12.0.1
multidict==5.1.0
oauthlib==3.1.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pykube-ng==20.10.0
pyrsistent==0.17.3
python-dateutil==2.8.1
python-json-logger==2.0.1
PyYAML==5.3.1
redis==3.5.3
requests==2.25.0
requests-oauthlib==1.3.0
rsa==4.6
six==1.15.0
typing-extensions==3.7.4.3
urllib3==1.26.2
websocket-client==0.57.0
yarl==1.6.3
